Added use_java9 boolean flag for optional JDK9 usage, to be used in cheets_CTS_R
By default the flag is False. The flag will be turned on in cheets_CTS_R.py.
The path modification should be per process, so it will not affect other tests that are not cheets_CTS_R, unless the flag is turned on.

BUG=b:142092636
TEST=Tried path change commands in a python shell of same LXC base image.

Change-Id: I520ecd5f4ab2f2bf797feafb14a564fa08a53a49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2082419
Tested-by: Jiyoun Ha <jiyounha@google.com>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Jiyoun Ha <jiyounha@google.com>
Auto-Submit: Jiyoun Ha <jiyounha@google.com>