[autotest] reduce timeouts when reaching out to nonexistent service AFE
The plan is to first land this change on master, and then cherry-pick it
to release-80 if appropriate. On master, reducing the timeouts here is a
harmless no-op, but on the release 80 branch this change will save a lot
of time.

BUG=chromium:1058095
TEST=CQ

Change-Id: I9026c2c87f3848c4a342cf967ed33a5d3859e638
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2087110
Reviewed-by: Allen Li <ayatane@chromium.org>
Tested-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Gregory Nisbet <gregorynisbet@google.com>