autotest: Move platform_MemCheck to bvt-perbuild.
The test has been ported to Tast and is running as a part of the CQ.
It is, however, still used as a part of memory qual sequence, so we want
to be able to track it. We move to to bvt-perbuild to reduce resource
consumption.

BUG=chromium:1028902
TEST=None

Change-Id: Idc86c9ad44866d17bdf960db72438d62e8250686
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2097780
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>