faft-cr50: temporarily remove firmware_Cr50InvalidateRW
Remove this test until the issues are figured out. The long runtime is
causing the suite to timeout before all tests can run. Remove the suite
from the control files, so it wont get scheduled. Control file changes
need an autotest push to take effect. Also skip in run_once, so the main
part of the test wont run while we're waiting for the autotest push.

This should be reverted after the bug is fixed.

BUG=b:151473821
TEST=none

Change-Id: Iabdc4dbd454b6c2205d8afe2d60bb2166edc30ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2103530
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>