bluetooth: test_discover_by_device should poll
The test_discovery_by_device triggers an active scan and can be flaky if
the scan isn't performed long enough or the peer's advertising interval
is low. Make the test poll discovery for 60s instead, which is the same
behavior the controller has when looking for the peer.

BUG=b:151634876
TEST=Ran bluetooth_AdapterSRSanity

Change-Id: I31b0aa0a9ccb41363ff75237159526d455d6decb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2105917
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Auto-Submit: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>