Fix au_fsi tests for moblab
The au_fsi tests fail on moblab due to a permissions issue in the
directory created by omaha devserver. The directory is mounted as
noexec, so the gsutil installed by the devserver is not able to run
properly to stage the payload for the tests.

BUG=b:150283761
TEST=None

Change-Id: I745fa20bac5640454e1d11675a159e5aad51a26c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2111640
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>