autotest: Use existing exception for last log line.
The previous rework of this code changed the exception type raised,
causing failures.  Now the last log line will be injected into the
body of the exception, but the exception type is preserved, to avoid
interfering with code that expects socket.error or BadStatusLine.

BUG=b:152168940
TEST=Run firmware_FAFTRPC with DUT python2 chmod -x.

Change-Id: I0b85151604589c4fb780c6b3dfc8c3c1b0ddc3bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2116711
Tested-by: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>