Parse update_engine logs to retrieve hostlogs
The production devservers do not accumulate the hostlogs for AU tests by
default. We used to gather these hostlogs only for devserver instances
that we started along with lab devservers. But we don't do that anymore
and now there was no way to get those hostlogs. But, the parameters that
we are looking for do kind of exist in the update_engine logs
themselves. So now instead of trying to get those hostlogs from
devservers, we extract them from the update_engine logs themselves.

This will also allow us to deprecate the hostlog API from devservers.

BUG=chromium:1064135
TEST=test_that chromeos6-row4-rack10-host1.cros autoupdate_ForcedOOBEUpdate.full -b reef --no-retry --autotest_dir /home/ahassani/trunk/src/third_party/autotest/files --ssh_private_key /home/ahassani/.ssh/testing_rsa --args="job_repo_url=http://chromeos6-devserver4.cros.corp.google.com:8082/static/reef-release/R83-12996.0.0/autotest/packages"

Change-Id: Ibf6dc3f0b323511b9cf945883f2ff797b76c815b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2120287
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>