servo_host: grab all servod logs since restart
Servod might crash in the middle of a run for any number of reasons and
we need to debug that. This change caches the instance's log timestamp
on init, and compares it to the timestamp at the end. If they are not
the same, it will find all timestamps (i.e. instances that have run
before but were turned down/crashed) between the init timestamp and the
current one, and grab those logs as well.

BUG=chromium:1056780
TEST=test_that --autotest_dir . $DUT servo_LogGrab --args "servo_host=$LABS"
     // test is added in subsequent CL

Change-Id: I60ee7c00a757ed3f9dc2ac87d5c2f0a567266d14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2121366
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>